refactor(backend): Split monolithic GraphQL resolvers into domain-based modules
## Overview
Refactored the 1521-line `resolvers.ts` file into 21 modular files organized by domain to minimize merge conflicts and improve maintainability.

## Changes

### Structure
- Created `graphql/index.ts` as main schema export point
- Split resolvers into 7 domains: sessions, queue, board, climbs, users, favorites, ticks
- Extracted shared utilities: helpers, async-iterators, types

### Files Created (21)
- `resolvers/shared/` - helpers.ts, async-iterators.ts, types.ts
- `resolvers/sessions/` - queries.ts, mutations.ts, subscriptions.ts, type-resolvers.ts
- `resolvers/queue/` - mutations.ts, subscriptions.ts, type-resolvers.ts
- `resolvers/board/` - queries.ts
- `resolvers/climbs/` - queries.ts, field-resolvers.ts
- `resolvers/users/` - queries.ts, mutations.ts
- `resolvers/favorites/` - queries.ts, mutations.ts
- `resolvers/ticks/` - queries.ts, mutations.ts
- `resolvers/index.ts` - combines all domain resolvers

### Files Modified (2)
- `graphql/yoga.ts` - updated schema import path
- `websocket/setup.ts` - updated schema import path

### Files Deleted (1)
- `graphql/resolvers.ts` - old monolithic file

## Bug Fixes
- Fixed missing `isActive` property in `mySessions` query

## Benefits
1. **Merge Conflict Prevention** - Team can work on different domains simultaneously
2. **Smaller Files** - Average ~80 lines vs 1521 in monolith
3. **Clear Ownership** - Well-defined domain boundaries
4. **Better Testing** - Domains can be tested in isolation
5. **Improved DX** - Faster navigation and IDE performance

## Testing
- ✅ TypeScript compilation passes
- ✅ All resolver logic preserved (no breaking changes)
- ✅ Zero new errors introduced
