+++ packages/backend/src/graphql/resolvers/queue/subscriptions.ts
@@ -1,0 +1,49 @@
+import type { ConnectionContext, QueueEvent } from '@boardsesh/shared-schema';
+import { roomManager } from '../../../services/room-manager.js';
+import { pubsub } from '../../../pubsub/index.js';
+import { requireSessionMember } from '../shared/helpers.js';
+import { createEagerAsyncIterator } from '../shared/async-iterators.js';
+
+export const queueSubscriptions = {
+  /**
+   * Subscribe to real-time queue updates
+   * Sends initial FullSync event followed by incremental updates
+   * Uses eager subscription to prevent race conditions
+   */
+  queueUpdates: {
+    subscribe: async function* (_: unknown, { sessionId }: { sessionId: string }, ctx: ConnectionContext) {
+      // Verify user is a member of the session they're subscribing to
+      // Uses retry logic to handle race conditions with joinSession
+      await requireSessionMember(ctx, sessionId);
+
+      // IMPORTANT: Subscribe to pubsub FIRST, before fetching state.
+      // This prevents a race condition where events could be published
+      // between fetching the queue state and starting to listen.
+      // Events that arrive before we yield FullSync will be queued.
+      // NOTE: We await here to ensure Redis subscription is established
+      // before proceeding - this is critical for multi-instance sync.
+      const asyncIterator = await createEagerAsyncIterator<QueueEvent>((push) => {
+        return pubsub.subscribeQueue(sessionId, push);
+      });
+
+      // Now fetch the current state (any events during this time are queued)
+      const queueState = await roomManager.getQueueState(sessionId);
+
+      // Send initial FullSync
+      yield {
+        queueUpdates: {
+          __typename: 'FullSync',
+          state: queueState,
+        } as QueueEvent,
+      };
+
+      // Continue with queued and new events
+      // Note: The client's reducer handles duplicate events via UUID deduplication,
+      // so any events that occurred before FullSync but are also in FullSync
+      // will be safely ignored.
+      for await (const event of asyncIterator) {
+        yield { queueUpdates: event };
+      }
+    },
+  },
+};
