+++ packages/backend/src/graphql/resolvers/queue/type-resolvers.ts
@@ -1,0 +1,11 @@
+import type { QueueEvent } from '@boardsesh/shared-schema';
+
+/**
+ * Union type resolver for QueueEvent
+ * GraphQL needs to know which concrete type to return
+ */
+export const queueEventResolver = {
+  __resolveType(obj: QueueEvent) {
+    return obj.__typename;
+  },
+};
