+++ packages/backend/src/graphql/resolvers/ticks/mutations.ts
@@ -1,0 +1,77 @@
+import { v4 as uuidv4 } from 'uuid';
+import type { ConnectionContext } from '@boardsesh/shared-schema';
+import { db } from '../../../db/client.js';
+import * as dbSchema from '@boardsesh/db/schema';
+import { requireAuthenticated, validateInput } from '../shared/helpers.js';
+import { SaveTickInputSchema } from '../../../validation/schemas.js';
+
+export const tickMutations = {
+  /**
+   * Save a tick (climb attempt/ascent) for the authenticated user
+   */
+  saveTick: async (
+    _: unknown,
+    { input }: { input: unknown },
+    ctx: ConnectionContext
+  ): Promise<unknown> => {
+    requireAuthenticated(ctx);
+
+    // Validate input with business rules
+    const validatedInput = validateInput(SaveTickInputSchema, input, 'input');
+
+    const userId = ctx.userId!;
+    const uuid = uuidv4();
+    const now = new Date().toISOString();
+    const climbedAt = new Date(validatedInput.climbedAt).toISOString();
+
+    // Insert into database
+    const [tick] = await db
+      .insert(dbSchema.boardseshTicks)
+      .values({
+        uuid,
+        userId,
+        boardType: validatedInput.boardType,
+        climbUuid: validatedInput.climbUuid,
+        angle: validatedInput.angle,
+        isMirror: validatedInput.isMirror,
+        status: validatedInput.status,
+        attemptCount: validatedInput.attemptCount,
+        quality: validatedInput.quality ?? null,
+        difficulty: validatedInput.difficulty ?? null,
+        isBenchmark: validatedInput.isBenchmark,
+        comment: validatedInput.comment,
+        climbedAt,
+        createdAt: now,
+        updatedAt: now,
+        sessionId: validatedInput.sessionId ?? null,
+        // Aurora sync fields are null - will be populated by periodic sync job
+        auroraType: null,
+        auroraId: null,
+        auroraSyncedAt: null,
+        auroraSyncError: null,
+      })
+      .returning();
+
+    return {
+      uuid: tick.uuid,
+      userId: tick.userId,
+      boardType: tick.boardType,
+      climbUuid: tick.climbUuid,
+      angle: tick.angle,
+      isMirror: tick.isMirror,
+      status: tick.status,
+      attemptCount: tick.attemptCount,
+      quality: tick.quality,
+      difficulty: tick.difficulty,
+      isBenchmark: tick.isBenchmark,
+      comment: tick.comment,
+      climbedAt: tick.climbedAt,
+      createdAt: tick.createdAt,
+      updatedAt: tick.updatedAt,
+      sessionId: tick.sessionId,
+      auroraType: tick.auroraType,
+      auroraId: tick.auroraId,
+      auroraSyncedAt: tick.auroraSyncedAt,
+    };
+  },
+};
